refactor: extract modbus value conversion into modbusValues.ts
fileToModbusSpecification, copyModbusDataToEntity, copyFromTestData,
setIdentifiedByEntities plus the ImodbusValues types move out of the
M2mSpecification god object as free functions; the class keeps one-line
static delegates for its published API. index.ts re-exports the new
module, so all existing imports resolve unchanged.
