+++ backend/src/specification/index.ts
@@ -3,6 +3,7 @@ export * from './converter.js'
 export * from './log.js'
 export * from './ifilespecification.js'
 export * from './m2mspecification.js'
+export * from './modbusValues.js'
 export * from './m2mgithub.js'
 export * from './configspec.js'
 export * from './migrator.js'
